Various UI fixes and Bootstrap 5 migrations
Fix for change detection error when opening visual dialog

Fix for Figures spacing and info tooltip position

Added translations for publication / citation modal footer text

Changed missing Boostrap 5 spacing migrations

Notification banner & Bootstrap migrations

ui fixes

Mobile spacing fixes

Disable style bleeding from notification banner

Funding call mobile layout

Funding call mobile spacing

Shared button styles

Disable scroll when main navigation is open
