+++ src/app/layout/header/header.component.scss
@@ -10,7 +10,7 @@
   position: fixed;
   z-index: 999;
   left: 0;
-  top: 78px;
+  top: 88px;
   background: rgba(0, 0, 0, 0.5);
   width: 100%;
   height: 100vh;
@@ -27,6 +27,7 @@ app-header {
   }
 
   .navbar-brand {
+    display: block;
     margin-right: 0;
   }
 
@@ -208,7 +209,7 @@ app-header {
   .collapse.navbar-collapse.menu.mobile-bar {
     position: absolute;
     right: -110vw;
-    top: 44px;
+    top: 66px;
     z-index: 1000;
     padding: 10px 8px 40px 8px;
     transition: right 0.25s ease-out;
