+++ src/app/portal/components/search-bar/search-bar.component.scss
@@ -23,7 +23,7 @@ app-search-bar {
   button.btn {
     border-left: 0;
     font-size: 1rem;
-    padding: 0 1.5rem;
+    padding: 0.5rem 1.5rem;
   }
 
   button.btn:focus {
@@ -312,16 +312,18 @@ app-search-bar {
     border-radius: 0.25rem;
   }
 
-  // Overlay
+  // Overlay base style comes from header component style
   #overlay {
-    position: absolute;
     z-index: 98;
-    left: 0;
-    background: rgba(0, 0, 0, 0.5);
-    width: 100%;
   }
 
   // Responsive
+  @media (max-width: 1199px) {
+    #overlay {
+      top: 66px;
+    }
+  }
+
   @media (max-width: 1054px) {
     .search-help {
       display: none;
@@ -337,6 +339,15 @@ app-search-bar {
   @media (max-width: 768px) {
     .search {
       padding: 30px 0 !important;
+    }
+
+    .has-search .form-control {
+      border-top-left-radius: 0.25rem !important;
+      border-bottom-left-radius: 0.25rem !important;
+    }
+
+    input#searchInput {
+      border-radius: 0.25rem !important;
     }
 
     .no-heading.search {
@@ -361,7 +372,7 @@ app-search-bar {
     }
 
     .input-group-append {
-      margin-left: 0.8rem;
+      margin-left: 0.8rem !important;
     }
 
     // Target
@@ -399,7 +410,7 @@ app-search-bar {
     }
 
     button.btn.btn-search {
-      padding: 0 1rem;
+      padding: 0.5rem 1rem;
     }
   }
 }
