+++ src/app/shared/components/buttons/buttons-shared-theme.scss
@@ -1,0 +1,47 @@
+// This file is part of the research.fi API service
+//
+// Copyright 2019 Ministry of Education and Culture, Finland
+//
+// :author: CSC - IT Center for Science Ltd., Espoo Finland servicedesk@csc.fi
+// :license: MIT
+
+@import '@angular/material/theming';
+
+@mixin buttons-shared-theme($theme) {
+  $primary: map-get($theme, primary);
+  $warn: map-get($theme, warn);
+
+  .primary-action,
+  .secondary-button {
+    border: 2px solid mat-color($primary);
+  }
+
+  .primary-action {
+    background-color: mat-color($primary);
+    color: mat-contrast($primary, 500);
+
+    &:hover,
+    &:active {
+      &:not([disabled]) {
+        border-color: mat-contrast($primary, 500);
+      }
+    }
+  }
+
+  .secondary-button {
+    background-color: mat-contrast($primary, 500);
+    color: mat-color($primary);
+
+    &:hover,
+    &:active {
+      &:not([disabled]) {
+        background-color: mat-color($primary, 100);
+      }
+    }
+  }
+
+  .close-button {
+    color: mat-color($primary, 500);
+    background-color: mat-contrast($primary, 500);
+  }
+}
