+++ src/app/shared/components/buttons/primary-action-button/primary-action-button.component.ts
@@ -2,19 +2,20 @@ import { Component, OnInit, Input } from '@angular/core';
 import { IconDefinition } from '@fortawesome/free-solid-svg-icons';
 
 @Component({
-  selector: 'app-secondary-button',
-  templateUrl: './secondary-button.component.html',
-  styleUrls: ['./secondary-button.component.scss'],
+  selector: 'app-primary-action-button',
+  templateUrl: './primary-action-button.component.html',
+  styleUrls: ['../buttons-shared.scss'],
 })
-export class SecondaryButtonComponent implements OnInit {
+export class PrimaryActionButtonComponent implements OnInit {
   @Input() icon: IconDefinition;
   @Input() iconAfter: boolean;
-  @Input() iconOnly: boolean;
   @Input() noSmIcon: boolean;
   @Input() content: string;
   @Input() disabled: boolean;
   @Input() big: boolean;
   @Input() small: boolean;
+  @Input() disabledRounded: boolean;
+  @Input() betaButton: boolean;
   @Input() px0: boolean;
 
   constructor() {}
