--- src/app/shared/components/close-button/close-button.component.scss
@@ -1,19 +1,0 @@
-// # This file is part of the research.fi API service
-// #
-// # Copyright 2019 Ministry of Education and Culture, Finland
-// #
-// # :author: CSC - IT Center for Science Ltd., Espoo Finland servicedesk@csc.fi
-// # :license: MIT
-
-.close-button {
-  font-weight: bold;
-  border: none;
-
-  &:hover {
-    text-decoration: underline;
-  }
-
-  fa-icon {
-    padding-right: 10px;
-  }
-}
