--- src/app/shared/components/primary-action-button/primary-action-button.component.scss-theme.scss
@@ -1,27 +1,0 @@
-// # This file is part of the research.fi API service
-// #
-// # Copyright 2019 Ministry of Education and Culture, Finland
-// #
-// # :author: CSC - IT Center for Science Ltd., Espoo Finland servicedesk@csc.fi
-// # :license: MIT
-
-@import '@angular/material/theming';
-
-@mixin app-primary-action-button-theme($theme) {
-  $primary: map-get($theme, primary);
-  $warn: map-get($theme, warn);
-
-  .primary-action {
-    background-color: mat-color($primary);
-
-    border: 2px solid mat-color($primary);
-    color: mat-contrast($primary, 500);
-
-    &:hover,
-    &:active {
-      &:not([disabled]) {
-        border-color: mat-contrast($primary, 500);
-      }
-    }
-  }
-}
