--- src/app/shared/components/secondary-button/secondary-button.component.scss
@@ -1,62 +1,0 @@
-// # This file is part of the research.fi API service
-// #
-// # Copyright 2019 Ministry of Education and Culture, Finland
-// #
-// # :author: CSC - IT Center for Science Ltd., Espoo Finland servicedesk@csc.fi
-// # :license: MIT
-
-.secondary-button {
-  &:active {
-    box-shadow: 0 0 0 4px;
-  }
-
-  &:disabled {
-    opacity: 0.6;
-  }
-
-  &.icon:not(.icon-after) {
-    padding-left: 0;
-  }
-
-  &.icon.icon-after {
-    padding-right: 0;
-  }
-
-  &.icon.icon-only {
-    padding: 0;
-    & fa-icon {
-      padding: 1rem;
-    }
-  }
-
-  &.big {
-    width: 100%;
-    padding-top: 10px;
-    padding-bottom: 10px;
-  }
-
-  &.small {
-    padding: 0 10px;
-  }
-
-  fa-icon.before {
-    padding: 0 16px 0 24px;
-  }
-
-  fa-icon.after {
-    padding: 0 24px 0 16px;
-  }
-}
-
-@media (max-width: 575px) {
-  .secondary-button {
-    &.no-sm-icon {
-      padding-left: unset;
-      padding-right: unset;
-    }
-
-    fa-icon.d-none {
-      padding: 0;
-    }
-  }
-}
