--- src/app/shared/components/secondary-button/secondary-button.component.scss-theme.scss
@@ -1,27 +1,0 @@
-// # This file is part of the research.fi API service
-// #
-// # Copyright 2019 Ministry of Education and Culture, Finland
-// #
-// # :author: CSC - IT Center for Science Ltd., Espoo Finland servicedesk@csc.fi
-// # :license: MIT
-
-@import '@angular/material/theming';
-
-@mixin app-secondary-button-theme($theme) {
-  $primary: map-get($theme, primary);
-  $warn: map-get($theme, warn);
-
-  .secondary-button {
-    background-color: mat-contrast($primary, 500);
-
-    border: 2px solid mat-color($primary);
-    color: mat-color($primary);
-
-    &:hover,
-    &:active {
-      &:not([disabled]) {
-        background-color: mat-color($primary, 100);
-      }
-    }
-  }
-}
